EOS-27348: Refactor py-utils code for cortx confstore
