+++ py-utils/src/utils/common/common.py
@@ -81,35 +81,3 @@ class CortxConf:
         if CortxConf._cluster_conf is None:
             raise ConfError(errno.ENOENT, "Path for config file, cluster.conf, not provided")
         return CortxConf._cluster_conf
-
-
-class ConfigStore:
-
-    """ CORTX Config Store """
-
-    _conf_idx = "cortx_conf"
-
-    def __init__(self, conf_url):
-        """ Initialize with the CONF URL."""
-        self._conf_url = conf_url
-        Conf.load(self._conf_idx, self._conf_url, skip_reload=True)
-
-    def set_kvs(self, kvs: list):
-        """
-        Parameters:
-        kvs - List of KV tuple, e.g. [('k1','v1'),('k2','v2')]
-        """
-
-        for key, val in kvs:
-            Conf.set(self._conf_idx, key, val)
-        Conf.save(self._conf_idx)
-
-    def set(self, key: str, val: str):
-        """Sets the value into conf in-memory at the given key."""
-        Conf.set(self._conf_idx, key, val)
-        Conf.save(self._conf_idx)
-
-    def get(self, key: str) -> str:
-        """ Returns value for the given key """
-
-        return Conf.get(self._conf_idx, key)
